# inventarioBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Inventario API",
        "description": "Sistema para gestionar inventario para empresas en crecimiento",
        "contact": {
            "name": "API Support",
            "email": "support@inventario.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "description": "Obtiene una lista de todos los clientes registrados en el sistema",
                "operationId": "f477bf009248c3386340b4ca00d97efe",
                "responses": {
                    "200": {
                        "description": "Lista de clientes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron clientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear nuevo cliente",
                "description": "Crea un nuevo cliente en el sistema",
                "operationId": "eac5506362a5880581b92359106f2612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "identificacion",
                                    "telefono",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "María García"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@ejemplo.com"
                                    },
                                    "identificacion": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/clientes/{id}": {
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "description": "Actualiza la información de un cliente existente",
                "operationId": "04a60664b93f6d2476bfda4f083b4c19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "María García"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@ejemplo.com"
                                    },
                                    "identificacion": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "description": "Elimina un cliente del sistema",
                "operationId": "df4292d1f4d1b1beb6ac4b0ab8a444cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El cliente se ha eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/clientes/{id}/estado": {
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cambiar estado del cliente",
                "description": "Alterna el estado (activo/inactivo) de un cliente",
                "operationId": "5f6bf42cb67be496d4b7bfa39bee823c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del cliente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estado del cliente ha sido actualizado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
        "/clientes/buscar": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Buscar cliente por nombre",
                "description": "Busca clientes por nombre (búsqueda parcial)",
                "operationId": "564a7e06a0f0e491639d0b52e3f79e38",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clientes encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetro nombre requerido"
                    },
                    "404": {
                        "description": "No se encontraron clientes"
                    }
                }
            }
        },
        "/clientes/activos": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes activos",
                "description": "Obtiene una lista de todos los clientes activos",
                "operationId": "824f7269846d9047c9440991c0662fec",
                "responses": {
                    "200": {
                        "description": "Lista de clientes activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay clientes activos"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "María García"
                    },
                    "identificacion": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "email": {
                        "type": "string",
                        "example": "maria@ejemplo.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "name": "bearerAuth",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}